extract avatar container component
