+++ teaching-assistant/src/app/demo/_views/Demo.tsx
@@ -1,67 +1,33 @@
 "use client";
 
-import { useState, useEffect, useRef, useCallback } from "react";
-import {
-  DemoSidebar,
-  ConversationPopup,
-  ConversationTracker,
-} from "@/components";
+import { useState, useEffect, useCallback, useRef } from "react";
+import { DemoSidebar, ConversationTracker } from "@/components";
 import { useAnamContext, useSettingsContext } from "@/contexts";
-import { errorHandler, logger } from "@/utils";
-import {
-  AnamEvent,
-  Message,
-  MessageRole,
-} from "@anam-ai/js-sdk/dist/module/types";
+import { AnamEvent, Message, MessageRole } from "@anam-ai/js-sdk/dist/module/types";
 import { useRouter } from "next/navigation";
-import { Spinner } from "@radix-ui/themes";
+import { AvatarContainer } from "@/components/AvatarContainer";
 
 const INITIAL_TIME_LEFT = 120;
 
 export function DemoView() {
-  const { anamClient, isClientInitialized } = useAnamContext();
+  const { anamClient } = useAnamContext();
   const { selectedLanguage } = useSettingsContext();
   const router = useRouter();
 
-  const videoRef = useRef<HTMLVideoElement | null>(null);
-  const audioRef = useRef<HTMLAudioElement | null>(null);
-  const streamingStartedRef = useRef(false);
-  const listenersAddedRef = useRef(false);
-
-  const [isVideoStreaming, setIsVideoStreaming] = useState(false);
-  const [loadingText, setLoadingText] = useState("Connecting...");
-  const [streamError, setStreamError] = useState<string | null>(null);
   const [conversation, setConversation] = useState<
     { sender: string; text: string }[]
   >([]);
   const [timeLeft, setTimeLeft] = useState(INITIAL_TIME_LEFT);
 
+  const listenersAddedRef = useRef(false);
+
   const stopStreaming = useCallback(() => {
     if (anamClient) {
       anamClient.stopStreaming().catch((error) => {
-        errorHandler(error);
         console.error("Failed to stop streaming:", error);
       });
     }
   }, [anamClient]);
-
-  const onConnectionEstablished = useCallback(() => {
-    setLoadingText("Connected to a Persona...");
-    setStreamError(null);
-    logger.info("Connection established");
-  }, []);
-
-  const onVideoStartedStreaming = useCallback(() => {
-    setLoadingText("");
-    setStreamError(null);
-    setIsVideoStreaming(true);
-    logger.info("Video started streaming");
-  }, []);
-
-  const onConnectionClosed = useCallback((reason: string) => {
-    logger.info("Connection closed", reason);
-    setIsVideoStreaming(false);
-  }, []);
 
   const updateConversation = useCallback((updatedMessages: Message[]) => {
     const mappedMessages = updatedMessages.map((message) => ({
@@ -74,35 +40,23 @@ export function DemoView() {
     setConversation(mappedMessages);
   }, []);
 
-  // Handle the timer
+  const handleStreamingEnd = useCallback(() => {
+    router.push("/");
+  }, [router]);
+
   useEffect(() => {
     if (timeLeft === 0) {
-      if (streamingStartedRef.current) {
-        stopStreaming();
-        streamingStartedRef.current = false;
-      }
-      router.push("/");
+      stopStreaming();
     }
     const timer =
-      isVideoStreaming && timeLeft > 0
-        ? setInterval(() => setTimeLeft((prev) => prev - 1), 1000)
-        : undefined;
+      timeLeft > 0 ? setInterval(() => setTimeLeft((prev) => prev - 1), 1000) : undefined;
     return () => {
       if (timer) clearInterval(timer);
     };
-  }, [isVideoStreaming, timeLeft, stopStreaming, router]);
+  }, [timeLeft, stopStreaming]);
 
   useEffect(() => {
     if (anamClient && !listenersAddedRef.current) {
-      anamClient.addListener(
-        AnamEvent.CONNECTION_ESTABLISHED,
-        onConnectionEstablished
-      );
-      anamClient.addListener(
-        AnamEvent.VIDEO_PLAY_STARTED,
-        onVideoStartedStreaming
-      );
-      anamClient.addListener(AnamEvent.CONNECTION_CLOSED, onConnectionClosed);
       anamClient.addListener(
         AnamEvent.MESSAGE_HISTORY_UPDATED,
         updateConversation
@@ -113,18 +67,6 @@ export function DemoView() {
     return () => {
       if (anamClient && listenersAddedRef.current) {
         anamClient.removeListener(
-          AnamEvent.CONNECTION_ESTABLISHED,
-          onConnectionEstablished
-        );
-        anamClient.removeListener(
-          AnamEvent.VIDEO_PLAY_STARTED,
-          onVideoStartedStreaming
-        );
-        anamClient.removeListener(
-          AnamEvent.CONNECTION_CLOSED,
-          onConnectionClosed
-        );
-        anamClient.removeListener(
           AnamEvent.MESSAGE_HISTORY_UPDATED,
           updateConversation
         );
@@ -132,49 +74,7 @@ export function DemoView() {
         listenersAddedRef.current = false;
       }
     };
-  }, [
-    anamClient,
-    onConnectionEstablished,
-    onVideoStartedStreaming,
-    onConnectionClosed,
-    updateConversation,
-  ]);
-
-  useEffect(() => {
-    const startStreaming = async () => {
-      if (
-        !isClientInitialized ||
-        !anamClient ||
-        !videoRef.current ||
-        !audioRef.current ||
-        streamingStartedRef.current
-      )
-        return;
-
-      try {
-        await anamClient.streamToVideoAndAudioElements(
-          videoRef.current.id,
-          audioRef.current.id
-        );
-        streamingStartedRef.current = true;
-      } catch (error) {
-        streamingStartedRef.current = false;
-        errorHandler(error);
-        const message =
-          error instanceof Error ? error.message : "An unknown error occurred";
-        setStreamError(message);
-      }
-    };
-
-    startStreaming();
-
-    return () => {
-      if (streamingStartedRef.current) {
-        stopStreaming();
-        streamingStartedRef.current = false;
-      }
-    };
-  }, [isClientInitialized, anamClient, videoRef, audioRef, stopStreaming]);
+  }, [anamClient, updateConversation]);
 
   const progressValue =
     ((INITIAL_TIME_LEFT - timeLeft) / INITIAL_TIME_LEFT) * 100;
@@ -185,31 +85,10 @@ export function DemoView() {
         {/* Main Content */}
         <div className="flex flex-col items-center p-5 flex-1">
           <div className="relative">
-            <div className="relative rounded-lg overflow-hidden">
-              <video
-                id="avatar-video"
-                ref={videoRef}
-                autoPlay
-                playsInline
-                className="rounded-lg object-cover w-[90vw] h-auto md:w-auto md:h-[90vh]"
-              />
-              <audio id="avatar-audio" ref={audioRef} autoPlay hidden />
-              <div className="flex justify-center items-center w-full h-full absolute top-0 left-0">
-                {streamError ? (
-                  <p className="text-red-500 text-base">{streamError}</p>
-                ) : loadingText ? (
-                  <div className="flex flex-col items-center">
-                    <Spinner size="3" />
-                  </div>
-                ) : null}
-              </div>
-              {/* Conversation Popup */}
-              {isVideoStreaming && (
-                <div className="absolute bottom-4 right-4 hidden md:flex">
-                  <ConversationPopup conversation={conversation} />
-                </div>
-              )}
-            </div>
+            <AvatarContainer
+              conversation={conversation}
+              onStreamingEnd={handleStreamingEnd}
+            />
           </div>
           {/* Progress Bar */}
           <div className="w-full max-w-2xl mt-4">
